dashboard/requirements.txt
streamlit==1.32.2
tensorflow-cpu==2.13.0
gdown
numpy
pillow
opencv-python-headless
keras==2.15.0